requirements.txt
# Python dependencies for Liquid Glass Music Player
# Install with: pip install -r requirements.txt

# Audio metadata extraction
mutagen>=1.47.0

# Image processing for album art
Pillow>=9.0.0

# Optional: For better audio format support
# eyed3>=0.9.7  # Additional MP3 support
# tinytag>=1.8.1  # Lightweight metadata reader

# Version: v5.2.0
